센터 상세페이지 기능 추가
센터 위치 지도에 표시 및 modal창 생성

+++ src/main/java/edu/pnu/domain/Board.java
@@ -24,7 +24,7 @@ public class Board {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long seq;
-	private int center_id;
+	private Long center_id;
 	private String content;
 	
 	@Temporal(TemporalType.TIMESTAMP)
